fuzzing: more revew related cleaning
