+++ Cargo.toml
@@ -33,6 +33,3 @@ libp2p = { version ="0.39", features = ["mdns", "tcp-tokio"]}
 parking_lot = "0.11"
 env_logger = "0.8"
 chrono = "0.4"
-
-[features]
-mock_common = []
