+++ src/lib.rs
@@ -8,7 +8,7 @@ use std::{fmt::Debug, hash::Hash as StdHash, pin::Pin};
 
 use crate::nodes::NodeMap;
 
-pub use config::{default_config, Config, DelayConfig};
+pub use config::{default_config, exponential_slowdown, Config, DelayConfig};
 pub use member::Member;
 pub use network::{Network, NetworkData};
 pub use nodes::{NodeCount, NodeIndex};
@@ -25,7 +25,7 @@ pub use signed::*;
 mod config;
 pub mod rmc;
 mod terminal;
-#[cfg(any(test, feature = "mock_common"))]
+#[cfg(test)]
 pub mod testing;
 mod units;
