+++ src/testing/mock/Cargo.toml
@@ -6,7 +6,7 @@ publish = false
 edition = "2018"
 
 [dependencies]
-aleph-bft = { path = "../../..", features = ["mock_common"] }
+aleph-bft = { path = "../../.." }
 async-trait = "0.1"
 codec = { package = "parity-scale-codec", version = "2", default-features = false, features = ["std"] }
 env_logger = { version = "0.8", default-features = false }
