--- src/testing/mock_common.rs
@@ -1,26 +1,0 @@
-use crate::{
-    config::{exponential_slowdown, Config, DelayConfig},
-    NodeCount, NodeIndex,
-};
-
-use std::{sync::Arc, time::Duration};
-
-pub fn gen_config(node_ix: NodeIndex, n_members: NodeCount) -> Config {
-    let delay_config = DelayConfig {
-        tick_interval: Duration::from_millis(5),
-        requests_interval: Duration::from_millis(50),
-        unit_broadcast_delay: Arc::new(|t| exponential_slowdown(t, 100.0, 1, 3.0)),
-        //100, 100, 300, 900, 2700, ...
-        unit_creation_delay: Arc::new(|t| exponential_slowdown(t, 50.0, usize::MAX, 1.000)),
-        //50, 50, 50, 50, ...
-    };
-    Config {
-        node_ix,
-        session_id: 0,
-        n_members,
-        delay_config,
-        rounds_margin: 200,
-        max_units_per_alert: 200,
-        max_round: 5000,
-    }
-}
