+++ src/testing/mod.rs
@@ -10,8 +10,6 @@ mod crash;
 mod dag;
 #[cfg(test)]
 pub(crate) mod mock;
-#[cfg(any(test, feature = "mock_common"))]
-pub mod mock_common;
 #[cfg(test)]
 mod rmc;
 #[cfg(test)]
